Card keys must be 32-bit.

+++ src/SevenEval.h
@@ -20,6 +20,7 @@
 #ifndef SKPOKEREVAL_SEVENEVAL_H
 #define SKPOKEREVAL_SEVENEVAL_H
 
+#include <cstdint>
 #include "Constants.h"
 
 class SevenEval {
@@ -31,12 +32,12 @@ public:
   // The higher the rank the better the hand. Two hands of equal rank tie.
   short unsigned GetRank(int, int, int, int, int, int, int) const;
 private:
-  short unsigned *mRankPtr;
-  short unsigned *mFlushRankPtr;
-  long unsigned mDeckcardsKey[DECK_SIZE];
-  short unsigned mDeckcardsFlush[DECK_SIZE];
-  short unsigned mDeckcardsSuit[DECK_SIZE];
-  short mFlushCheck[MAX_FLUSH_CHECK_SUM+1];
+  uint16_t *mRankPtr;
+  uint16_t *mFlushRankPtr;
+  uint32_t mDeckcardsKey[DECK_SIZE];
+  uint16_t mDeckcardsFlush[DECK_SIZE];
+  uint16_t mDeckcardsSuit[DECK_SIZE];
+  int16_t mFlushCheck[MAX_FLUSH_CHECK_SUM+1];
 };
 
 #endif // SKPOKEREVAL_SEVENEVAL_H
